move to record type NHSNumber value and hide constructor
This is a step towards constraining the construction of an NHSNumber type to only valid NHS numbers.  A simple `List Int` type alias
did not prevent invariants from being assigned to an NHSNumber (e.g. 9 integers instead of 10).  The module no longer exports the constructor, so all instantiation of an NHSNumber now requires a valid number at time of creation.